requirements.txt
appdirs==1.4.3
backports.ssl-match-hostname==3.5.0.1
certifi==2017.4.17
chardet==3.0.3
funcsigs==1.0.2
idna==2.5
mock==2.0.0
packaging==16.8
pbr==3.0.1
pyparsing==2.2.0
PyYAML==3.12
redis==2.10.5
requests==2.17.3
six==1.10.0
slackclient==1.0.5
urllib3==1.21.1
websocket-client==0.40.0
